refactor(blorktools): panels to heading
