+++ packages/blorktools/src/asset_debugger/panels/asset-panel/asset-panel.html
@@ -26,8 +26,8 @@
             <span class="collapse-indicator">[+]</span>
         </div>
         <div class="metadata-content section-content" style="display: none;">
-            <!-- Atlas Panel -->
-            <div id="atlas-panel-container" class="panel-container">
+            <!-- Atlas Heading -->
+            <div id="atlas-heading-container" class="panel-container">
                 <div class="texture-type-selector">
                     <button data-texture-type="baseColor" class="texture-type-button active">Base Color</button>
                     <button data-texture-type="orm" class="texture-type-button">ORM</button>
@@ -49,8 +49,8 @@
             <span class="collapse-indicator">[+]</span>
         </div>
         <div class="metadata-content section-content" style="display: none;">
-            <!-- UV Panel -->
-            <div id="uv-panel-container" class="panel-container">
+            <!-- UV Heading -->
+            <div id="uv-heading-container" class="panel-container">
                 <div id="uv-info-container" class="uv-info-container">
                     <p>No UV data loaded yet.</p>
                 </div>
@@ -92,8 +92,8 @@
             <span class="collapse-indicator">[+]</span>
         </div>
         <div class="metadata-content section-content" style="display: none;">
-            <!-- Rig Panel -->
-            <div id="rig-panel-container" class="panel-container">
+            <!-- Rig Heading -->
+            <div id="rig-heading-container" class="panel-container">
                 <div id="rig-content" class="rig-container">
                     <p>No rig data loaded yet.</p>
                 </div>
