+++ packages/blorktools/src/asset_debugger/panels/asset-panel/asset-panel.js
@@ -4,10 +4,10 @@
  * This module handles sample sections and their collapsible functionality.
  */
 import { getState } from '../../util/state/scene-state';
-import { createMeshVisibilityPanel } from '../mesh-panel/mesh-panel';
-import { initAtlasPanel, updateAtlasVisualization } from '../atlas-panel/atlas-panel';
-import { initUvPanel, updateUvPanel } from '../uv-panel/uv-panel.js';
-import { updateRigPanel } from '../rig-panel/rig-panel';
+import { createMeshVisibilityPanel } from './mesh-heading/mesh-heading';
+import { initAtlasPanel, updateAtlasVisualization } from './atlas-heading/atlas-heading';
+import { initUvPanel, updateUvPanel } from './uv-heading/uv-heading.js';
+import { updateRigPanel } from './rig-heading/rig-heading';
 
 // Track initialization state
 let controlsInitialized = false;
