+++ packages/blorktools/src/asset_debugger/panels/asset-panel/uv-heading/uv-heading.css
