--- packages/blorktools/src/asset_debugger/panels/atlas-panel/atlas-panel.html
@@ -1,17 +1,0 @@
-<!-- Atlas Panel HTML Component -->
-<div id="atlas-panel-container" class="tab-content-inner">
-    <h3>Atlas Visualization</h3>
-    <div class="texture-type-selector">
-        <button data-texture-type="baseColor" class="texture-type-button active">🎨 Base Color</button>
-        <button data-texture-type="orm" class="texture-type-button">✨ ORM</button>
-        <button data-texture-type="normal" class="texture-type-button">🧩 Normal</button>
-    </div>
-    
-    <!-- Atlas texture container -->
-    <div id="basecolor-container" class="texture-preview">
-        <canvas id="atlas-canvas" class="atlas-display" width="260" height="260"></canvas>
-        <div id="atlas-label" class="texture-label">Base Color Atlas</div>
-    </div>
-    
-    <div id="segment-info" class="segment-info" style="display: none;"></div>
-</div>
