--- packages/blorktools/src/asset_debugger/panels/mesh-panel/mesh-panel.html
@@ -1,9 +1,0 @@
-<!-- Mesh Panel Component -->
-<div id="mesh-tab" class="tab-content">
-    <h3>Mesh Visibility</h3>
-    <div id="mesh-groups"></div>
-    
-    <div class="button-container" style="margin-top: 20px; text-align: center;" data-has-meshes="false">
-        <button id="download-asset-btn" class="theme-toggle" style="display: none;">Download Asset</button>
-    </div>
-</div>
